Add decryption/decryptingtrustee remote webapps. decryption app does not know the trustee directory.
