+++ decryptingtrustee/src/main/kotlin/electionguard/webapps/decryptingtrustee/plugins/Serialization.kt
@@ -1,0 +1,11 @@
+package electionguard.webapps.decryptingtrustee.plugins
+
+import io.ktor.serialization.kotlinx.json.*
+import io.ktor.server.plugins.contentnegotiation.*
+import io.ktor.server.application.*
+
+fun Application.configureSerialization() {
+    install(ContentNegotiation) {
+        json()
+    }
+}
