--- decryptingtrustee/src/main/kotlin/webapps/electionguard/Application.kt
@@ -1,35 +1,0 @@
-package webapps.electionguard
-
-import electionguard.core.PowRadixOption
-import electionguard.core.ProductionMode
-import electionguard.core.productionGroup
-import io.ktor.server.application.*
-import io.ktor.server.config.*
-import io.ktor.server.engine.*
-import io.ktor.server.plugins.callloging.*
-import io.ktor.server.request.*
-import org.slf4j.event.Level
-import webapps.electionguard.plugins.*
-
-val groupContext = productionGroup(PowRadixOption.HIGH_MEMORY_USE, ProductionMode.Mode4096)
-
-fun main(args: Array<String>) {
-    io.ktor.server.netty.EngineMain.main(args)
-}
-
-@Suppress("unused") // application.conf references the main function. This annotation prevents the IDE from marking it as unused.
-fun Application.module() {
-    install(CallLogging) {
-        level = Level.INFO
-        format { call ->
-            val status = call.response.status()
-            val httpMethod = call.request.httpMethod.value
-            val path = call.request.path()
-            "Status: $status, HTTP method: $httpMethod, Path: $path"
-        }
-    }
-    configureSecurity()
-    configureAdministration()
-    configureSerialization()
-    configureRouting()
-}
