--- decryptingtrustee/src/main/kotlin/webapps/electionguard/plugins/Serialization.kt
@@ -1,11 +1,0 @@
-package webapps.electionguard.plugins
-
-import io.ktor.serialization.kotlinx.json.*
-import io.ktor.server.plugins.contentnegotiation.*
-import io.ktor.server.application.*
-
-fun Application.configureSerialization() {
-    install(ContentNegotiation) {
-        json()
-    }
-}
