+++ decryptingtrustee/src/main/resources/application.conf
@@ -7,8 +7,3 @@ ktor {
         modules = [ webapps.electionguard.ApplicationKt.module ]
     }
 }
-jwt {
-    domain = "https://jwt-provider-domain/"
-    audience = "jwt-audience"
-    realm = "ktor sample app"
-}
