--- decryption/src/main/kotlin/webapps/electionguard/Application.kt
@@ -1,20 +1,0 @@
-package webapps.electionguard
-
-import electionguard.core.PowRadixOption
-import electionguard.core.ProductionMode
-import electionguard.core.productionGroup
-import io.ktor.server.application.*
-import webapps.electionguard.plugins.*
-
-val groupContext = productionGroup(PowRadixOption.HIGH_MEMORY_USE, ProductionMode.Mode4096)
-
-fun main(args: Array<String>): Unit =
-    io.ktor.server.netty.EngineMain.main(args)
-
-@Suppress("unused") // application.conf references the main function. This annotation prevents the IDE from marking it as unused.
-fun Application.module() {
-    configureSecurity()
-    configureSerialization()
-    configureAdministration()
-    configureRouting()
-}
