--- decryption/src/main/kotlin/webapps/electionguard/plugins/Routing.kt
@@ -1,9 +1,0 @@
-package webapps.electionguard.plugins
-
-import io.ktor.server.routing.*
-import io.ktor.server.application.*
-
-fun Application.configureRouting() {
-    routing {
-    }
-}
