--- decryption/src/main/kotlin/webapps/electionguard/plugins/Security.kt
@@ -1,7 +1,0 @@
-package webapps.electionguard.plugins
-
-import io.ktor.server.application.*
-
-fun Application.configureSecurity() {
-
-}
