--- decryption/src/main/resources/application.conf
@@ -1,9 +1,0 @@
-ktor {
-    deployment {
-        port = 11191
-        port = ${?PORT}
-    }
-    application {
-        modules = [ webapps.electionguard.ApplicationKt.module ]
-    }
-}
