+++ settings.gradle.kts
@@ -111,4 +111,6 @@ include ("encryptserver")
 include ("encryptclient")
 include ("keyceremonytrustee")
 include ("keyceremony")
+include ("decryptingtrustee")
+include ("decryption")
