+++ testInput/chained/private_data/trustees/decryptingTrustee-guardian1.json
@@ -1,0 +1,10 @@
+{
+    "id": "guardian1",
+    "x_coordinate": 1,
+    "polynomial_coefficients": [
+        "0A301C7020D55FE6A04F87AC19DB3A6209F5A9BDFA5DCD9E630AE4F4F98D2C55",
+        "2B126948FA109327EF6E6EEE634EA64555566A006D885699DBF4CC30DA8A544C",
+        "37108C6F8A257E73FDC14CA3CBEF9F841D2FB9633AFD531D6DE9A54FD009274C"
+    ],
+    "key_share": "3A961963DC247485562034FB6C46DA5073C6F45A9A147437FA087C684AA04E83"
+}
