+++ testInput/chained/private_data/trustees/decryptingTrustee-guardian2.json
@@ -1,0 +1,10 @@
+{
+    "id": "guardian2",
+    "x_coordinate": 2,
+    "polynomial_coefficients": [
+        "B3727CA099D7AD6FDFC11DD0E58DA7679074ECED36F8F92F5E14260D9C63762D",
+        "66180558FC1767EF16A45EA8E3827AE0ADA4C44D0DCBB28E4AAC5F03574CBB89",
+        "2EBB80791178D5D381264F49E4AC4C5E0AE69699BAF21782514871E83A9B22A8"
+    ],
+    "key_share": "4AB5023B838E65E4D1A6A6C76FAFBEEE2D31ED00B93238C3363C9AE7D5769E4E"
+}
