+++ testInput/chained/private_data/trustees/decryptingTrustee-guardian3.json
@@ -1,0 +1,10 @@
+{
+    "id": "guardian3",
+    "x_coordinate": 3,
+    "polynomial_coefficients": [
+        "B89EE1B03885BFF9F51811655688FC78081A379F34A3C5A2F7E9D1832ADA9E81",
+        "0D450AE6938D87088D9D36A858F9E6A15A328E77C343F03D1CC4355FD77D5AAD",
+        "C0191831C39DD0CDCE5FDDEBC5EE08654BFE194DFF9283C23E68281675DC56D3"
+    ],
+    "key_share": "A69E3547E970A16EE7BC0C46602C8C1ACEC5B83CC353DA126DA53804614E3121"
+}
