+++ testInput/unchained/private_data/trustees/decryptingTrustee-guardian1.json
@@ -1,0 +1,10 @@
+{
+    "id": "guardian1",
+    "x_coordinate": 1,
+    "polynomial_coefficients": [
+        "D55A716AA7A8FE921D8C53A3D3E9DD0BD7C932EE8510F0E4AA7DEFADEB628FF0",
+        "A78940AF20E75B5967A663BD10C67458F34E88A6E7A2E6AD643438D726CCF2C5",
+        "E89AB9948EF00480362C9CF6454A782E80FDA89FB0415033B1E76D01D7C5FA60"
+    ],
+    "key_share": "B1641F7343693910D4595B20FB19438409C37E6CD21F1931C89C4EAAA9224A40"
+}
