+++ testInput/unchained/private_data/trustees/decryptingTrustee-guardian2.json
@@ -1,0 +1,10 @@
+{
+    "id": "guardian2",
+    "x_coordinate": 2,
+    "polynomial_coefficients": [
+        "0581B3016FB32911BA648D48DAE69D97E8B9CA00483217DF84A114181406153E",
+        "D0726398A51F05E53FF9F580088C65E57916BE80C40D2B050CC14EE8C0399B68",
+        "A3B39C21C9329E2BF44664CE6AFCCB1B016FAFBA232F939FFD98BBD960945454"
+    ],
+    "key_share": "D71A055730DE4BC230DD3829A3BF8BE24B2DFF04C236FEB18BE343AEF708B10F"
+}
