+++ testInput/unchained/private_data/trustees/decryptingTrustee-guardian3.json
@@ -1,0 +1,10 @@
+{
+    "id": "guardian3",
+    "x_coordinate": 3,
+    "polynomial_coefficients": [
+        "D90A687923BF4449203EC75C4D02A7E864A4BC94C877E07D137B24CDFC5FD320",
+        "7165C49B4FB3C91E745C44829143330649CF1FEA36DA4376766A36DFF60CF6C1",
+        "87CDD3F49A71001A95BA1353A468D069ABFA057D8668F6F3EF223E9B97EBFA9F"
+    ],
+    "key_share": "25083E91037AA4010DBB3F62F5C5FBA6E9673B4B660299C08C6F07A0E57BAE35"
+}
